Rename flow _block_ modules accordingly
