+++ src/oemof/solph/_groupings.py
@@ -24,10 +24,10 @@ SPDX-License-Identifier: MIT
 
 from oemof.network import groupings as groupings
 
-from oemof.solph.flows._investment_flow import InvestmentFlowBlock
-from oemof.solph.flows._non_convex_flow import NonConvexFlowBlock
-from oemof.solph.flows._non_convex_invest_flow import NonConvexInvestFlowBlock
-from oemof.solph.flows._simple_flow import SimpleFlowBlock
+from oemof.solph.flows._investment_flow_block import InvestmentFlowBlock
+from oemof.solph.flows._non_convex_flow_block import NonConvexFlowBlock
+from oemof.solph.flows._non_convex_invest_flow_block import NonConvexInvestFlowBlock
+from oemof.solph.flows._simple_flow_block import SimpleFlowBlock
 
 
 def constraint_grouping(node, fallback=lambda *xs, **ks: None):
