+++ src/oemof/solph/_models.py
@@ -23,10 +23,10 @@ from pyomo.opt import SolverFactory
 from oemof.solph import processing
 from oemof.solph.buses._bus import BusBlock
 from oemof.solph.components._transformer import TransformerBlock
-from oemof.solph.flows._investment_flow import InvestmentFlowBlock
-from oemof.solph.flows._non_convex_flow import NonConvexFlowBlock
-from oemof.solph.flows._non_convex_invest_flow import NonConvexInvestFlowBlock
-from oemof.solph.flows._simple_flow import SimpleFlowBlock
+from oemof.solph.flows._investment_flow_block import InvestmentFlowBlock
+from oemof.solph.flows._non_convex_flow_block import NonConvexFlowBlock
+from oemof.solph.flows._non_convex_invest_flow_block import NonConvexInvestFlowBlock
+from oemof.solph.flows._simple_flow_block import SimpleFlowBlock
 
 
 class LoggingError(BaseException):
