+++ src/oemof/solph/flows/_investment_flow_block.py
