+++ src/oemof/solph/flows/_non_convex_flow_block.py
