+++ src/oemof/solph/flows/_non_convex_invest_flow_block.py
