+++ src/oemof/solph/flows/_simple_flow_block.py
